+++ lib/list_base.dart
@@ -2,50 +2,118 @@
 // https://github.com/akserg/monomer
 // All rights reserved.  Please see the LICENSE.md file.
 
-library monomer_list_base;
+library monomer_listbase;
 
 import 'dart:html';
 import 'package:polymer/polymer.dart';
+import 'package:template_binding/template_binding.dart';
 
-import 'list_item_renderer.dart';
 import 'component.dart';
+import 'list_item_renderer.dart';
+import 'item_renderer_owner.dart';
+
 import 'utility.dart';
 
 /**
- *  The ListBase class is the base class for controls that represent lists
- *  of items that can have one or more selected and can scroll through the
- *  items.  Items are supplied using the <code>dataProvider</code> property
- *  and displayed via item renderers.
+ * The ListBase class is the base class for all components that support selection.
  */
-abstract class ListBase {
+@CustomTag('m-list-base')
+class ListBase extends DivElement with Polymer, Observable, Component implements ItemRendererOwner {
 
   /***********
    * CONSTANTS
    **********/
-  
-  static const String CHANGE_EVENT = "change";
+  /**
+   * Provider of 'change' events.
+   */
+  static const EventStreamProvider<Event> _changeEvent = const EventStreamProvider<Event>(Component.CHANGE_EVENT);
   
   /**************
-   * Properties *
+   * PROPERTIES *
    **************/
   
   /**
    * Data Provider instance
    */
-  ObservableList get dataProvider;
+  @published
+  ObservableList dataProvider = toObservable([]);
   
-  bool get allowSelectFirst;
-  bool get allowMultipleSelection;
-  bool get autoScrollToSelection;
-  String get labelPath;
-  String get valuePath;
+  /**
+   * Return an array of references to the selected items in the data provider.
+   */
+  @published
+  ObservableList selectedItems = toObservable([]);
+  
+  /**
+   * Instance of [ListItemRenderer] using to rendering content of items.
+   */
+  @published
+  String itemRenderer;
+  itemRendererChanged(old) {
+    callLater(updateUI);
+  }
+  
+  /**
+   * Map of NodeBindings.
+   */
+  Map<ListItemRenderer, NodeBindExtension> itemRenderers = {};
+  
+  /**
+   * Owner of ItemRenderers. By default it's [shadowRoot], but any other
+   * point in shadowRoot can be returned as well.
+   */
+  Node get owner => shadowRoot;
+  
+  /**
+   * Flag allows select first item by default.
+   */
+  @published
+  bool allowSelectFirst = false;
+  
+  /**
+   * Flag allows select multiple items. 
+   */
+  @published
+  bool allowMultipleSelection = false;
+  
+  /**
+   * Flag allows scroll to show selected items.
+   */
+  @published
+  bool autoScrollToSelection = false;
+  
+  /**
+   * Path used to find the label.
+   */
+  @published
+  String labelPath;
+  
+  /**
+   * Path used to find the label.
+   */
+  @published
+  Function labelFunction;
+  
+  /**
+   * Path used to find the value.
+   */
+  @published
+  String valuePath;
 
-  dynamic _value;
+  /**
+   * Separator used to split data before send to string
+   */
   String valueSeparator;
+
+  /**
+   * Inteesentation of [Value].
+   */
+  dynamic _value;
   
   /**
    * Return a reference to the selected item in the data provider.
    */
+  @published
   dynamic get selectedItem {
     print('get selectedItem');
     if (selectedItems.length > 0)
@@ -62,13 +130,8 @@ abstract class ListBase {
     if (item != null) {
       selectedItems.add(item);
     }
-    updateUI();
+    callLater(updateSelectedItems);
   }
-  
-  /**
-   * Return an array of references to the selected items in the data provider.
-   */
-  ObservableList get selectedItems;
   
   /**
    * Return the index in the data provider of the selected item.
@@ -83,7 +146,7 @@ abstract class ListBase {
   void set selectedIndex(int value) {
     if (value >= 0 && value < dataProvider.length) {
       selectedItem = dataProvider[value];
-      updateUI();
+      callLater(updateSelectedItems);
     } else if (value == -1) {
       selectedItem = null;
     } else {
@@ -114,7 +177,7 @@ abstract class ListBase {
         return _value;
       }
       if (valuePath != null) {
-        s = s[valuePath];
+        s = Utility.getValue(s, valuePath);
       }
       return s;
     }
@@ -130,67 +193,69 @@ abstract class ListBase {
       var dataItems = dataProvider;
       List result;
       if (allowMultipleSelection && valueSeparator != null) {
-        if (v is String) {
-          result = v.split(valueSeparator);
-        } else {
-          result = v; // as List
-        }
+        result = v.toString().split(valueSeparator);
       } else if (v is List) {
         result = v;
       } else {
         result = [v];
       }
       selectedItems.addAll(Utility.intersect(dataItems, valuePath, result));
-      updateUI();
+      callLater(updateSelectedItems);
     }
   }
   
-  /***********
-   * Methods *
-   ***********/
-
+  /**********
+   * Events *
+   **********/
+  
   /**
-   * Initialise ListBase mixin with [owner] info.
+   * Stream of 'change' events handled by this element.
    */
-  void initializeListBase(owner) {
-    onPropertyChange(owner, #dataProvider, (){
+  ElementStream<Event> get onChange => _changeEvent.forElement(this);
+  
+  /******************
+   * Initialisation *
+   ******************/
+  
+  /**
+   * Default factory constructor.
+   */
+  factory ListBase() {
+    return new Element.tag('div', 'm-list-base');
+  }
+  
+  /**
+   * Constructor instantiated by the DOM when a ListBase element has been 
+   * created.
+   */
+  ListBase.created() : super.created() {
+    onPropertyChange(this, #dataProvider, (){
       print('onPropertyChange dataProvider: $dataProvider');
+      // Remove selection and call updateSelectedItems methods afterwords
       selectAll(false);
+      callLater(updateUI);
     });
-    onPropertyChange(owner, #selectedItems, (){
+    onPropertyChange(this, #selectedItems, (){
       print('onPropertyChange selectedItems: $selectedItems');
-      owner.dispatchEvent(new CustomEvent(CHANGE_EVENT, detail:value));
+      callLater((){
+        updateSelectedItems();
+        dispatchEvent(new CustomEvent(Component.CHANGE_EVENT, detail:value));
+      });
     });
-    selectedItems.changes.listen((v) {
-      print('selectedItems changed: $selectedItems');
-      owner.dispatchEvent(new CustomEvent(CHANGE_EVENT, detail:value));
-    });
-  }
-
-  /**
-   * Return list of instanses of item renderer components.
-   */
-  List<Element> get itemRenderers;
-
-  /**
-   * Convinience method to update state of each item renderer. 
-   */
-  void updateUI() {
-    for (int i = 0; i < itemRenderers.length; i++) {
-      updateItem(itemRenderers[i]);
-    }
   }
   
-  /**
-   * Update state of [item].
-   */
-  void updateItem(Element item) {
-    if (item is ListItemRenderer && item is Component) {
-      (item as ListItemRenderer).itemSelected = isSelected((item as Component).data);
-    }
+  @override
+  void enteredView() {
+    super.enteredView();
+    print('enteredView');
+    callLater(updateUI);
   }
-
-   /**
+  
+  /***************
+   * Translation *
+   ***************/
+  
+  /**
    * Helper function return index of item [data].
    */
   int itemIndex(data) {
@@ -223,7 +288,8 @@ abstract class ListBase {
    */
   String itemToLabel(data) {
     print('itemToLabel for $data');
-    return labelPath == null ? data : data == null ? null : data[labelPath];
+    return labelFunction != null ? labelFunction(data) : 
+           labelPath == null ? data : data == null ? null : Utility.getValue(data, labelPath);
   }
   
   /**
@@ -231,9 +297,9 @@ abstract class ListBase {
    */
   String itemToValue(data) {
     print('itemToValue for $data');
-    return valuePath == null ? data : data == null ? null : data[valuePath];
+    return valuePath == null ? data : data == null ? null : Utility.getValue(data, valuePath);
   }
- 
+
   /****************************
    * Selection Business logic *
    ****************************/
@@ -254,7 +320,7 @@ abstract class ListBase {
     if (value && dataProvider != null) {
       selectedItems.addAll(dataProvider);
     }
-    updateUI();
+    callLater(updateSelectedItems);
   }
   
   
@@ -266,15 +332,13 @@ abstract class ListBase {
     if (allowSelectFirst && selectedIndex == 0)
       return;
     // Move through item renderers and select first selected
-    for (Iterator ae = itemRenderers.iterator; ae.moveNext();) {
-      Element item = ae.current;
-      if (item is ListItemRenderer) {
-        if ((item as ListItemRenderer).itemSelected) {
-          item.scrollIntoView();
-          return;
-        }
+    itemRenderers.keys.firstWhere((ListItemRenderer renderer){
+      if (renderer.itemSelected) {
+        renderer.scrollIntoView();
+        return true;
       }
-    }
+      return false;
+    });
   }
   
   /**
@@ -292,6 +356,88 @@ abstract class ListBase {
       selectedItems.clear();
       selectedItems.add(data);
     }
-    updateUI();
+    notifyPropertyChange(#selectedItems, null, selectedItems);
+    callLater(updateSelectedItems);
+  }
+  
+  /**********
+   * Update *
+   *********/
+  
+  /**
+   * Convinience method to update state of each item renderer. 
+   */
+  void updateSelectedItems() {
+    itemRenderers.forEach((ListItemRenderer renderer, NodeBindExtension nodeBind){
+      updateItem(renderer);
+    });
+  }
+  
+  /**
+   * Update state of [item].
+   */
+  void updateItem(ListItemRenderer item) {
+    item.itemSelected = isSelected(item.data);
+  }
+  
+  /**
+   * Update display list of UI.
+   */
+  void updateUI() {
+    print('updateUI');
+    removeUIItems();
+    addUIItems();
+  }
+  
+  /**
+   * Remove old ItemRenderers.
+   */
+  void removeUIItems() {
+    // Remove old
+    itemRenderers.forEach((ListItemRenderer item, NodeBindExtension bindExt) {
+      print('updateUI.remove $item');
+      bindExt.unbindAll();
+      item.remove();
+    });
+    itemRenderers.clear();
+  }
+  
+  /**
+   * Create new ItemRenderers based on [dataProvider] and [itemRenderer].
+   */
+  void addUIItems() {
+    if (itemRenderer != null) {
+      dataProvider.forEach((data){
+        //
+        print('updateUI.render $data');
+        // Instantiate [itemRenderer]
+        ListItemRenderer renderer = instantiateItemRenderer();
+        // Bind renderer properties to local
+        NodeBindExtension bindExt = nodeBind(renderer)
+            ..bind('data', data, '');
+        // Append new renderer to owner
+        owner.append(renderer);
+        // Keep them until update or remove
+        itemRenderers[renderer] = bindExt;
+        print('updateUI.add $renderer');
+      });
+    }
+  }
+  
+  /**
+   * Create an instance of [itemRenderer].
+   */
+  ListItemRenderer instantiateItemRenderer() {
+    ListItemRenderer renderer;
+    if (itemRenderer.contains('.')) {
+      List<String> parts = itemRenderer.split(".");
+      print('updateUI.itemRenderer are ${parts[0]}, ${parts[1]}');
+      renderer = new Element.tag(parts[0], parts[1]);
+    } else {
+      print('updateUI.itemRenderer is $itemRenderer');
+      renderer = new Element.tag(itemRenderer);
+    }
+    renderer.owner = this;
+    return renderer;
   }
 }