+++ lib/list_item_renderer.dart
@@ -4,34 +4,36 @@
 
 library monomer_list_item_renerer;
 
+import 'item_renderer.dart';
+
 /**
  * ListItemRenderer are dedicated to displaying a particular field
  * from the data provider item. 
  */
-abstract class ListItemRenderer {
+abstract class ListItemRenderer extends ItemRenderer {
   
-  /**
-   * Index of rendering data item.
-   * Sets by owner.
-   */
-  int itemIndex;
-  
-  /**
-   * Will equals true if rendering item is first. 
-   * Sets by owner.
-   */
-  bool itemFirst;
-  
-  /**
-   * Will equals true if rendering item is last.
-   * Sets by owner.
-   */
-  bool itemLast;
-  
-  /**
-   * Will equals true if item is odd. 
-   */
-  bool itemOdd = false;
+//  /**
+//   * Index of rendering data item.
+//   * Sets by owner.
+//   */
+//  int itemIndex;
+//  
+//  /**
+//   * Will equals true if rendering item is first. 
+//   * Sets by owner.
+//   */
+//  bool itemFirst;
+//  
+//  /**
+//   * Will equals true if rendering item is last.
+//   * Sets by owner.
+//   */
+//  bool itemLast;
+//  
+//  /**
+//   * Will equals true if item is odd. 
+//   */
+//  bool itemOdd;
   
   /**
    * Will equals true if rendering item is selected.