+++ lib/utility.dart
@@ -52,7 +52,7 @@ class Utility {
    * Find [value] in [valuePath] matching items in [list].
    */
   static dynamic getMatch(List list, String valuePath, dynamic value) {
-    var result = list.firstWhere((item){
+    return list.firstWhere((item){
       var dataItem = item;
       if (valuePath != null) {
         dataItem = getValue(item, valuePath);
@@ -62,7 +62,6 @@ class Utility {
       }
       return false;
     }, orElse:() => null);
-    return result;
   }
   
   /**
@@ -72,16 +71,30 @@ class Utility {
     assert(item != null);
     assert(valuePath != null);
     try {
-      InstanceMirror inst = reflect(item);
       int indx = valuePath.indexOf(".");
-      if (indx == -1) {
-        // We reached last chain in valuePath
-        InstanceMirror field = inst.getField(stringAsSymbol(valuePath));
-        return field == null ? null : field.reflectee;
+      if (item is Map) {
+        // This is Map - process key:value pairs
+        if (indx == -1) {
+          // Just get map element by key
+          return item[valuePath];
+        } else {
+          // Get chain
+          dynamic child = item[valuePath.substring(0, indx)];
+          // Process chain
+          return child == null ? null : getValue(child, valuePath.substring(indx + 1));
+        }
       } else {
-        InstanceMirror field = inst.getField(stringAsSymbol(valuePath.substring(0, indx)));
-        // we have other chains in valuePath - need recursion
-        return field == null ? null : getValue(field.reflectee, valuePath.substring(indx + 1));
+        // This is object - process by property name
+        InstanceMirror inst = reflect(item);
+        if (indx == -1) {
+          // We reached last chain in valuePath
+          InstanceMirror field = inst.getField(stringAsSymbol(valuePath));
+          return field == null ? null : field.reflectee;
+        } else {
+          InstanceMirror field = inst.getField(stringAsSymbol(valuePath.substring(0, indx)));
+          // we have other chains in valuePath - need recursion
+          return field == null ? null : getValue(field.reflectee, valuePath.substring(indx + 1));
+        }
       }
     } on Exception catch(e) {
       print(e);
@@ -96,15 +109,26 @@ class Utility {
     assert(item != null);
     assert(valuePath != null);
     try {
-      InstanceMirror inst = reflect(item);
       int indx = valuePath.indexOf(".");
-      if (indx == -1) {
-        // We reached last chain in valuePath
-        inst.setField(stringAsSymbol(valuePath), value);
+      if (item is Map) {
+        if (indx == -1) {
+          item[valuePath] = value;
+        } else {
+          // Get chain
+          dynamic child = item[valuePath.substring(0, indx)];
+          // Process chain
+          setValue(child, valuePath.substring(indx + 1), value);
+        }
       } else {
-        InstanceMirror field = inst.getField(stringAsSymbol(valuePath.substring(0, indx)));
-        // we have other chains in valuePath - need recursion
-        return field == null ? null : setValue(field.reflectee, valuePath.substring(indx + 1), value);
+        InstanceMirror inst = reflect(item);
+        if (indx == -1) {
+          // We reached last chain in valuePath
+          inst.setField(stringAsSymbol(valuePath), value);
+        } else {
+          InstanceMirror field = inst.getField(stringAsSymbol(valuePath.substring(0, indx)));
+          // we have other chains in valuePath - need recursion
+          setValue(field.reflectee, valuePath.substring(indx + 1), value);
+        }
       }
     } on Exception catch(e) {
       print(e);